requirements.txt
mss
Pillow
numpy
